Combine a database query with a cohort.

Combine a database query with a cohort, only retaining observations between time.prev days prior to indexdt, and time.post days after indexdt, and for test data with values between lower.bound and upper.bound. The most recent numobs observations will be returned. cohort must contain variables patid and indexdt. The type of query must be specified for appropriate data manipulation. Input type = med if interested in medical diagnoses from the observation file, and type = test if interseted in test data from the observation file.

Usage

combine_query(
  cohort,
  db.query,
  query.type = c("med", "test", "drugissue", "hes_primary", "death"),
  time.prev = Inf,
  time.post = Inf,
  lower.bound = -Inf,
  upper.bound = Inf,
  numobs = 1,
  value.na.rm = TRUE,
  earliest.values = FALSE,
  reduce.output = TRUE
)

Arguments

cohort

Cohort to combine with the database query.

db.query

Output from database query (ideally obtained through db_query).

query.type

Type of query

time.prev

Number of days prior to index date to look for codes.

time.post

Number of days after index date to look for codes.

lower.bound

Lower bound for returned values when query.type = "test".

upper.bound

Upper bound for returned values when query.type = "test".

numobs

Number of observations to be returned.

value.na.rm

If TRUE will remove data with NA in the value column of the queried data and remove values outside of lower.bound and upper.bound when query.type = "test".

earliest.values

If TRUE will return the earliest values as opposed to most recent.

reduce.output

If TRUE will reduce output to just patid, obsdate, medical/product code, and test value.

Value

A data.table with observations that meet specified criteria.

Details

value.na.rm = FALSE may be of use when extracting variables like smoking status, where we want test data for number of cigarettes per day, but do not want to remove all observations with NA in the value column, because the medcodeid itself may indicate smoking status.

Examples

## Create connection to a temporary database
aurum_extract <- connect_database(tempfile("temp.sqlite"))

## Add observation data from all observation files in specified directory
cprd_extract(db = aurum_extract,
filepath = system.file("aurum_data", package = "rAURUM"),
filetype = "observation")
#>   |                                                                              |                                                                      |   0%[1] "Adding /home/runner/work/_temp/Library/rAURUM/aurum_data/aurum_allpatid_set1_extract_observation_001.txt 2024-08-02 14:23:11.000272"
#>   |                                                                              |=======================                                               |  33%[1] "Adding /home/runner/work/_temp/Library/rAURUM/aurum_data/aurum_allpatid_set1_extract_observation_002.txt 2024-08-02 14:23:11.037861"
#>   |                                                                              |===============================================                       |  67%[1] "Adding /home/runner/work/_temp/Library/rAURUM/aurum_data/aurum_allpatid_set1_extract_observation_003.txt 2024-08-02 14:23:11.049202"
#>   |                                                                              |======================================================================| 100%

## Query database for a specific medcode
db.query <- db_query(db.open = aurum_extract,
tab ="observation",
codelist.vector = "187341000000114")

## Define cohort
pat<-extract_cohort(filepath = system.file("aurum_data", package = "rAURUM"))

### Add an index date to pat
pat$indexdt <- as.Date("01/01/2020", format = "%d/%m/%Y")

## Combine query with cohort retaining most recent three records
combine_query(cohort = pat,
db.query = db.query,
query.type = "med",
numobs = 3)
#>     patid       medcodeid obsdate
#>    <char>          <char>   <num>
#> 1:      1 187341000000114   -5373
#> 2:      2 187341000000114   -5769
#> 3:      6 187341000000114  -14727