Combine a database query with a cohort returning a 0/1 vector depending on whether each individual has a recorded code of interest.

Combine a database query with a cohort returning a 0/1 vector depending on whether each individual has a recorded code of interest. cohort must contain variables patid and indexdt. The database query will be merged with the cohort by variable patid. If an individual has at least numobs observations between time.prev days prior to indexdt, and time.post days after indexdt, a 1 will be returned, 0 otherwise. The type of query must be specified for appropriate data manipulation.

Usage

combine_query_boolean(
  cohort,
  db.query,
  query.type = c("observation", "drugissue", "hes_primary"),
  time.prev = Inf,
  time.post = 0,
  numobs = 1
)

Arguments

cohort

Cohort to combine with the database query.

db.query

Output from database query (ideally obtained through db_query).

query.type

Type of query

time.prev

Number of days prior to index date to look for codes.

time.post

Number of days after index date to look for codes.

numobs

Number of observations required to be observed in specified time window to return a 1.

Value

A 0/1 vector.

Examples

## Create connection to a temporary database
aurum_extract <- connect_database(tempfile("temp.sqlite"))

## Add observation data from all observation files in specified directory
cprd_extract(db = aurum_extract,
filepath = system.file("aurum_data", package = "rAURUM"),
filetype = "observation")
#>   |                                                                              |                                                                      |   0%[1] "Adding /home/runner/work/_temp/Library/rAURUM/aurum_data/aurum_allpatid_set1_extract_observation_001.txt 2024-07-26 15:36:29.852"
#>   |                                                                              |=======================                                               |  33%[1] "Adding /home/runner/work/_temp/Library/rAURUM/aurum_data/aurum_allpatid_set1_extract_observation_002.txt 2024-07-26 15:36:29.864411"
#>   |                                                                              |===============================================                       |  67%[1] "Adding /home/runner/work/_temp/Library/rAURUM/aurum_data/aurum_allpatid_set1_extract_observation_003.txt 2024-07-26 15:36:29.875749"
#>   |                                                                              |======================================================================| 100%

## Query database for a specific medcode
db.query <- db_query(db.open = aurum_extract,
tab ="observation",
codelist.vector = "187341000000114")

## Define cohort
pat<-extract_cohort(filepath = system.file("aurum_data", package = "rAURUM"))

### Add an index date to pat
pat$indexdt <- as.Date("01/01/2020", format = "%d/%m/%Y")

## Combine query with cohort creating a 'history of' boolean variable
combine_query_boolean(cohort = pat,
db.query = db.query,
query.type = "observation")
#>  [1] 1 1 0 0 0 1 0 0 0 0 0 0