Extract most recent total cholesterol/high-density lipoprotein ratio score relative to an index date.

Extract most recent total cholesterol/high-density lipoprotein ratio score relative to an index date.

Usage

extract_cholhdl_ratio(
  cohort,
  varname = NULL,
  codelist.ratio,
  codelist.chol,
  codelist.hdl,
  codelist.ratio.vector,
  codelist.chol.vector,
  codelist.hdl.vector,
  indexdt,
  t = NULL,
  t.varname = TRUE,
  time.prev = 365.25 * 5,
  time.post = 0,
  lower.bound = -Inf,
  upper.bound = Inf,
  db.open = NULL,
  db = NULL,
  db.filepath = NULL,
  out.save.disk = FALSE,
  out.subdir = NULL,
  out.filepath = NULL,
  return.output = TRUE
)

Arguments

cohort

Cohort to extract age for.

varname

Optional name for variable in output dataset.

codelist.ratio

Name of codelist (stored on hard disk in "codelists/analysis/") for ratio to query the database with.

codelist.chol

Name of codelist (stored on hard disk in "codelists/analysis/") for total cholesterol to query the database with.

codelist.hdl

Name of codelist (stored on hard disk in "codelists/analysis/") for high-density lipoprotein to query the database with.

codelist.ratio.vector

Vector of codes for ratio to query the database with.

codelist.chol.vector

Vector of codes for total cholesterol to query the database with.

codelist.hdl.vector

Vector of codes for high-density lipoprotein to query the database with.

indexdt

Name of variable which defines index date in cohort.

t

Number of days after index date at which to calculate variable.

t.varname

Whether to add t to varname.

time.prev

Number of days prior to index date to look for codes.

time.post

Number of days after index date to look for codes.

lower.bound

Lower bound for returned values.

upper.bound

Upper bound for returned values.

db.open

An open SQLite database connection created using RSQLite::dbConnect, to be queried.

db

Name of SQLITE database on hard disk (stored in "data/sql/"), to be queried.

db.filepath

Full filepath to SQLITE database on hard disk, to be queried.

out.save.disk

If TRUE will attempt to save outputted data frame to directory "data/extraction/".

out.subdir

Sub-directory of "data/extraction/" to save outputted data frame into.

out.filepath

Full filepath and filename to save outputted data frame into.

return.output

If TRUE will return outputted data frame into R workspace.

Value

A data frame with variable total cholesterol/high-density lipoprotein ratio.

Details

Specifying db requires a specific underlying directory structure. The SQLite database must be stored in "data/sql/" relative to the working directory. If the SQLite database is accessed through db, the connection will be opened and then closed after the query is complete. The same is true if the database is accessed through db.filepath. A connection to the SQLite database can also be opened manually using RSQLite::dbConnect, and then using the object as input to parameter db.open. After wards, the connection must be closed manually using RSQLite::dbDisconnect. If db.open is specified, this will take precedence over db or db.filepath.

If out.save.disk = TRUE, the data frame will automatically be written to an .rds file in a subdirectory "data/extraction/" of the working directory. This directory structure must be created in advance. out.subdir can be used to specify subdirectories within "data/extraction/". These options will use a default naming convetion. This can be overwritten using out.filepath to manually specify the location on the hard disk to save. Alternatively, return the data frame into the R workspace using return.output = TRUE and then save onto the hard disk manually.

Specifying the non-vector type codelists requires a specific underlying directory structure. The codelist on the hard disk must be stored in "codelists/analysis/" relative to the working directory, must be a .csv file, and contain a column "medcodeid", "prodcodeid" or "ICD10" depending on the chosen tab. The input to these variables should just be the name of the files (excluding the suffix .csv). The codelists can also be read in manually, and supplied as a character vector. This option will take precedence over the codelists stored on the hard disk if both are specified.

Examples

## Connect
aurum_extract <- connect_database(tempfile("temp.sqlite"))

## Create SQLite database using cprd_extract
cprd_extract(aurum_extract,
filepath = system.file("aurum_data", package = "rAURUM"),
filetype = "observation", use.set = FALSE)
#> [1] "Adding /home/runner/work/_temp/Library/rAURUM/aurum_data/aurum_allpatid_set1_extract_observation_001.txt 2024-06-17 09:49:31.555667"
#> [1] "Adding /home/runner/work/_temp/Library/rAURUM/aurum_data/aurum_allpatid_set1_extract_observation_002.txt 2024-06-17 09:49:31.568582"
#> [1] "Adding /home/runner/work/_temp/Library/rAURUM/aurum_data/aurum_allpatid_set1_extract_observation_003.txt 2024-06-17 09:49:31.579854"

## Define cohort and add index date
pat<-extract_cohort(system.file("aurum_data", package = "rAURUM"))
pat$indexdt <- as.Date("01/01/1955", format = "%d/%m/%Y")

## Extract most recent cholhdl_ratio prior to index date
extract_cholhdl_ratio(cohort = pat,
codelist.ratio.vector = "498521000006119",
codelist.chol.vector = "401539014",
codelist.hdl.vector = "13483031000006114",
indexdt = "indexdt",
time.prev = Inf,
db.open = aurum_extract,
return.output = TRUE)
#>    patid cholhdl_ratio
#> 1      1            48
#> 2     10            NA
#> 3     11            NA
#> 4     12            NA
#> 5      2            NA
#> 6      3            NA
#> 7      4            NA
#> 8      5            18
#> 9      6            NA
#> 10     7            NA
#> 11     8            NA
#> 12     9            NA