Extract impotence status.

Extract impotence status. This requires either a medical diagnosis, or drug.numobs number of prescriptions in drug.time.prev number of days prior to the index date.

Usage

extract_impotence(
  cohort,
  varname = NULL,
  codelist.med,
  codelist.drug,
  codelist.med.vector,
  codelist.drug.vector,
  indexdt,
  t = NULL,
  t.varname = TRUE,
  drug.time.prev = Inf,
  drug.numobs = 1,
  db.open = NULL,
  db = NULL,
  db.filepath = NULL,
  out.save.disk = FALSE,
  out.subdir = NULL,
  out.filepath = NULL,
  return.output = TRUE
)

Arguments

cohort

Cohort to extract age for.

varname

Optional name for variable in output dataset.

codelist.med

Name of codelist (stored on hard disk in "codelists/analysis/") for medical diagnoses to query the database with.

codelist.drug

Name of codelist (stored on hard disk in "codelists/analysis/") for prescriptions to query the database with.

codelist.med.vector

Vector of codes for medical diagnoses to query the database with.

codelist.drug.vector

Vector of codes for prescriptions to query the database with.

indexdt

Name of variable which defines index date in cohort.

t

Number of days after index date at which to calculate variable.

t.varname

Whether to add t to varname.

drug.time.prev

Number of days prior to index date in which to look for prescriptions.

drug.numobs

Numbre of prescriptions required in time period to indicate impotence.

db.open

An open SQLite database connection created using RSQLite::dbConnect, to be queried.

db

Name of SQLITE database on hard disk (stored in "data/sql/"), to be queried.

db.filepath

Full filepath to SQLITE database on hard disk, to be queried.

out.save.disk

If TRUE will attempt to save outputted data frame to directory "data/extraction/".

out.subdir

Sub-directory of "data/extraction/" to save outputted data frame into.

out.filepath

Full filepath and filename to save outputted data frame into.

return.output

If TRUE will return outputted data frame into R workspace.

Value

A data frame with variable impotence status.

Details

Specifying db requires a specific underlying directory structure. The SQLite database must be stored in "data/sql/" relative to the working directory. If the SQLite database is accessed through db, the connection will be opened and then closed after the query is complete. The same is true if the database is accessed through db.filepath. A connection to the SQLite database can also be opened manually using RSQLite::dbConnect, and then using the object as input to parameter db.open. After wards, the connection must be closed manually using RSQLite::dbDisconnect. If db.open is specified, this will take precedence over db or db.filepath.

If out.save.disk = TRUE, the data frame will automatically be written to an .rds file in a subdirectory "data/extraction/" of the working directory. This directory structure must be created in advance. out.subdir can be used to specify subdirectories within "data/extraction/". These options will use a default naming convetion. This can be overwritten using out.filepath to manually specify the location on the hard disk to save. Alternatively, return the data frame into the R workspace using return.output = TRUE and then save onto the hard disk manually.

Specifying the non-vector type codelists requires a specific underlying directory structure. The codelist on the hard disk must be stored in "codelists/analysis/" relative to the working directory, must be a .csv file, and contain a column "medcodeid", "prodcodeid" or "ICD10" depending on the chosen tab. The input to these variables should just be the name of the files (excluding the suffix .csv). The codelists can also be read in manually, and supplied as a character vector. This option will take precedence over the codelists stored on the hard disk if both are specified.