Extract test data.

Query an RSQLite database and return a data frame containing the most recent test result that meets specified criteria.

Usage

extract_test_data(
  cohort,
  varname = NULL,
  codelist = NULL,
  codelist.vector = NULL,
  indexdt,
  t = NULL,
  t.varname = TRUE,
  time.prev = 365.25 * 5,
  time.post = 0,
  lower.bound = -Inf,
  upper.bound = Inf,
  db.open = NULL,
  db = NULL,
  db.filepath = NULL,
  out.save.disk = FALSE,
  out.subdir = NULL,
  out.filepath = NULL,
  return.output = FALSE
)

Arguments

cohort

Cohort of individuals to extract the 'history of' variable for.

varname

Name of variable in the outputted data frame.

codelist

Name of codelist (stored on hard disk) to query the database with.

codelist.vector

Vector of codes to query the database with. This takes precedent over codelist if both are specified.

indexdt

Name of variable in cohort which specifies the index date. The extracted variable will be calculated relative to this.

t

Number of days after indexdt at which to extract the variable.

t.varname

Whether to alter the variable name in the outputted data frame to reflect t.

time.prev

Number of days prior to index date to look for codes.

time.post

Number of days after index date to look for codes.

lower.bound

Lower bound for returned values.

upper.bound

Upper bound for returned values.

db.open

An open SQLite database connection created using RSQLite::dbConnect, to be queried.

db

Name of SQLITE database on hard disk (stored in "data/sql/"), to be queried.

db.filepath

Full filepath to SQLITE database on hard disk, to be queried.

out.save.disk

If TRUE will attempt to save outputted data frame to directory "data/extraction/".

out.subdir

Sub-directory of "data/extraction/" to save outputted data frame into.

out.filepath

Full filepath and filename to save outputted data frame into.

return.output

If TRUE will return outputted data frame into R workspace.

Value

A data frame containing most recent test result that meets required criteria.

Details

Specifying db requires a specific underlying directory structure. The SQLite database must be stored in "data/sql/" relative to the working directory. If the SQLite database is accessed through db, the connection will be opened and then closed after the query is complete. The same is true if the database is accessed through db.filepath. A connection to the SQLite database can also be opened manually using RSQLite::dbConnect, and then using the object as input to parameter db.open. After wards, the connection must be closed manually using RSQLite::dbDisconnect. If db.open is specified, this will take precedence over db or db.filepath.

If out.save.disk = TRUE, the data frame will automatically be written to an .rds file in a subdirectory "data/extraction/" of the working directory. This directory structure must be created in advance. out.subdir can be used to specify subdirectories within "data/extraction/". These options will use a default naming convetion. This can be overwritten using out.filepath to manually specify the location on the hard disk to save. Alternatively, return the data frame into the R workspace using return.output = TRUE and then save onto the hard disk manually.

Codelists can be specified in two ways. The first is to read the codelist into R as a character vector and then specify through the argument codelist.vector. Codelists stored on the hard disk can also be referred to from the codelist argument, but require a specific underlying directory structure. The codelist on the hard disk must be stored in a directory called "codelists/analysis/" relative to the working directory. The codelist must be a .csv file, and contain a column "medcodeid", "prodcodeid" or "ICD10" depending on the input for argument tab. The input to argument codelist should just be a character string of the name of the files (excluding the suffix '.csv'). The codelist.vector option will take precedence over the codelist argument if both are specified.

Currently only returns most recent test result. This will be updated to return more than one most recent test result if specified.

Examples

## Connect
aurum_extract <- connect_database(tempfile("temp.sqlite"))

## Create SQLite database using cprd_extract
cprd_extract(aurum_extract,
filepath = system.file("aurum_data", package = "rAURUM"),
filetype = "observation", use.set = FALSE)
#>   |                                                                              |                                                                      |   0%[1] "Adding /home/runner/work/_temp/Library/rAURUM/aurum_data/aurum_allpatid_set1_extract_observation_001.txt 2024-07-16 08:34:41.435862"
#>   |                                                                              |=======================                                               |  33%[1] "Adding /home/runner/work/_temp/Library/rAURUM/aurum_data/aurum_allpatid_set1_extract_observation_002.txt 2024-07-16 08:34:41.449086"
#>   |                                                                              |===============================================                       |  67%[1] "Adding /home/runner/work/_temp/Library/rAURUM/aurum_data/aurum_allpatid_set1_extract_observation_003.txt 2024-07-16 08:34:41.460441"
#>   |                                                                              |======================================================================| 100%

## Define cohort and add index date
pat<-extract_cohort(system.file("aurum_data", package = "rAURUM"))
pat$indexdt <- as.Date("01/01/1955", format = "%d/%m/%Y")

## Extract most recent test value prior to index date
extract_test_data(pat,
codelist.vector = "187341000000114",
indexdt = "fup_start",
db.open = aurum_extract,
time.prev = Inf,
return.output = TRUE)
#>    patid value
#> 1      1    NA
#> 2     10    NA
#> 3     11    NA
#> 4     12    NA
#> 5      2    46
#> 6      3    NA
#> 7      4    NA
#> 8      5    NA
#> 9      6    28
#> 10     7    NA
#> 11     8    NA
#> 12     9    NA